Prototype replays

+++ src/net/blockjack/ld26/states/PlayState.as
@@ -11,6 +11,7 @@ package net.blockjack.ld26.states
 	import org.flixel.FlxGroup;
 	import org.flixel.FlxObject;
 	import org.flixel.FlxPoint;
+	import org.flixel.FlxSave;
 	import org.flixel.FlxSprite;
 	import org.flixel.FlxState;
 	import org.flixel.FlxTilemap;
@@ -23,7 +24,10 @@ package net.blockjack.ld26.states
 	public class PlayState extends FlxState
 	{
 		private var player:Player;
-		private var gibs:FlxEmitter
+		private var gibs:FlxEmitter;
+		
+		private var replay:Boolean;
+		private var replays:FlxGroup;
 		
 		private var enemies:FlxGroup;
 		private var projectiles:FlxGroup;
@@ -39,6 +43,9 @@ package net.blockjack.ld26.states
 		private const RestartPNG:Class;
 		private var restartText:FlxSprite;
 		
+		[Embed(source="../../../../../assets/gfx/ui/InfoPopups.png")]
+		private const InfoPopupsPNG:Class;
+		
 		private var ui:FlxGroup;
 		
 		override public function create():void {
@@ -46,22 +53,40 @@ package net.blockjack.ld26.states
 			
 			super.create();
 			
+			replay = true;
+			
 			createLevel();
 			var startPos:FlxPoint = level.getStartPoint();
 			createGibs();
 			createEnemies();
 			createPlayer(startPos);
+			createReplays(startPos);
 			createUI();
 			
 			setupLevel();
 			
-			FlxG.flash(Main.BACKGROUND_COLOR, 0.5);
+			FlxG.flash(Main.BACKGROUND_COLOR, Main.TRANSITION_SPEED);
 		}
 		
 		private function createPlayer(startPos:FlxPoint):void {
-			player = new Player(gibs);
-			player.x = startPos.x;
-			player.y = startPos.y;
+			if(!replay) {
+				player = new Player(gibs, true);
+				player.x = startPos.x;
+				player.y = startPos.y;
+			}
+		}
+		
+		private function createReplays(startPos:FlxPoint):void {
+			if (replay) {
+				replays = new FlxGroup();
+				for (var i:int = 0; i < 20; i++) {
+					var player:Player = new Player(gibs, true);
+					player.x = startPos.x;
+					player.y = startPos.y;
+					
+					replays.add(player);
+				}
+			}
 		}
 		
 		private function createLevel():void {
@@ -78,7 +103,7 @@ package net.blockjack.ld26.states
 			gibs.gravity = 420;
 			gibs.particleDrag = new FlxPoint(100, 100);
 			gibs.setRotation( 0, 0);
-			gibs.makeParticles(GibsPNG, 20, 0, true, 0.5);
+			gibs.makeParticles(GibsPNG, 500, 0, true, 0.5);
 		}
 		
 		private function createEnemies():void {
@@ -94,6 +119,12 @@ package net.blockjack.ld26.states
 			restartText.y = Main.SWF_HEIGHT / 2 - restartText.height / 2;
 			restartText.visible = false;
 			
+			var infoPopup:FlxSprite = new FlxSprite(0, 0);
+			infoPopup.loadGraphic(InfoPopupsPNG, true, false, Main.SWF_WIDTH, Main.SWF_HEIGHT);
+			infoPopup.frame = Registry.levelNum;
+			infoPopup.drawFrame(true);
+			
+			ui.add(infoPopup);
 			ui.add(restartText);
 		}
 		
@@ -101,7 +132,14 @@ package net.blockjack.ld26.states
 			add(level.getBackground());
 			add(tilemapLevel);
 			add(tilemapObjects);
-			add(player);
+			
+			if (replay) {
+				add(replays);
+			}
+			else {
+				add(player);
+			}
+			
 			add(gibs);
 			add(enemies);
 			add(projectiles);
@@ -119,21 +157,36 @@ package net.blockjack.ld26.states
 		private function checkCollisions():void {
 			FlxG.collide(enemies, tilemapLevel);
 			FlxG.collide(projectiles, tilemapLevel, projectileCollideWithLevel);
-			FlxG.collide(player, tilemapLevel);
-			FlxG.overlap(player, enemies, playerCollideWithEnemy);
-			FlxG.overlap(player, projectiles, playerCollideWithProjectile);
-			FlxG.collide(player, tilemapObjects);
 			FlxG.collide(gibs, tilemapLevel);
+			
+			if(replay) {
+				FlxG.collide(replays, tilemapLevel);
+				FlxG.overlap(replays, enemies, playerCollideWithEnemy);
+				FlxG.overlap(replays, projectiles, playerCollideWithProjectile);
+				FlxG.collide(replays, tilemapObjects);
+			}
+			else {
+				FlxG.collide(player, tilemapLevel);
+				FlxG.overlap(player, enemies, playerCollideWithEnemy);
+				FlxG.overlap(player, projectiles, playerCollideWithProjectile);
+				FlxG.collide(player, tilemapObjects);
+			}
 		}
 		
 		private function checkReset():void {
-			if (!player.alive && FlxG.keys.justPressed("SPACE")) {
-				FlxG.switchState(new PlayState());
+			if (FlxG.keys.justPressed("SPACE")) {
+				if(replay || !player.alive) {
+					FlxG.switchState(new PlayState());
+				}
+			}
+			
+			if (FlxG.keys.justPressed("ESCAPE")) {
+				FlxG.switchState(new MainMenuState());
 			}
 		}
 		
 		public function playerCollideWithSpikes(tile:FlxTile, player:Player):void {
-			killPlayer();
+			killPlayer(player);
 		}
 		
 		public function projectileCollideWithLevel(projectile:Bullet, tiles:FlxTilemap):void {
@@ -142,19 +195,26 @@ package net.blockjack.ld26.states
 		
 		public function playerCollideWithEnemy(player:Player, enemy:Enemy):void {
 			if(FlxCollision.pixelPerfectCheck(player, enemy)) {
-				killPlayer();
+				killPlayer(player);
 			}
 		}
 		
 		public function playerCollideWithProjectile(player:Player, projectile:Bullet):void {
-			if(FlxCollision.pixelPerfectCheck(player, projectile)) {
-				killPlayer();
+			if(projectile.alive && FlxCollision.pixelPerfectCheck(player, projectile)) {
+				killPlayer(player);
 				projectile.kill();
 			}
 		}
 		
 		public function playerCollideWithExit(tile:FlxTile, player:Player):void {
-			nextLevel();
+			player.exit();
+			
+			if (replay) {
+				endReplay();
+			}
+			else {
+				nextLevel();
+			}
 		}
 		
 		public function playerCollideWithSpring(tile:FlxTile, player:Player):void {
@@ -171,19 +231,44 @@ package net.blockjack.ld26.states
 		
 		private function nextLevel():void {
 			player.alive = false;
+			
 			Registry.levelNum++;
+			Registry.unlockedToLevelNum = Math.max(Registry.levelNum, Registry.unlockedToLevelNum);
+			saveData();
+			
 			FlxG.fade(Main.BACKGROUND_COLOR, 0.5, function():void { FlxG.switchState(new PlayState()); } );
 		}
 		
-		public function killPlayer():void {
-			restartText.visible = true;
+		private function endReplay():void {
+			FlxG.fade(Main.BACKGROUND_COLOR, 0.5, function():void { FlxG.switchState(new PlayState()); } );
+		}
+		
+		public function killPlayer(player:Player = null):void {
+			if (!player) {
+				player = this.player;
+			}
+			
 			player.kill();
+			
+			if(!replay) {
+				restartText.visible = true;
+			}
 		}
 		
 		public function createProjectile(position:FlxPoint, direction:uint):void {
 			var projectile:Bullet = projectiles.recycle(Bullet) as Bullet
 			projectile.reset(position.x, position.y);
 			projectile.setDirection(direction);
+		}
+		
+		private function saveData():void {
+			var save:FlxSave = Registry.save;
+			
+			save.data.version = Main.VERSION;
+			save.data.levelNum = Registry.levelNum;
+			save.data.unlockedToLevelNum = Registry.unlockedToLevelNum;
+			
+			save.flush();
 		}
 		
 		override public function destroy():void {
